Large document zooming: Stop view scroll up/down
Resolves an issue where zooming in/out would cause the view to scroll up
or down when editing a large document. This is fixed by making
ZoomControl.cpp take into account that vertical padding doesn't scale
with zoom.

Unfortunately, while this fixes the scrolling issue, a viewport that
moves up and down as the user zooms is still an issue. The difference
is that while previously, zooming caused the view to move significantly,
the view now remains roughly centered.